requirements.txt
dvc
dagshub
scikit-learn
mlflow
dvc-s3
python-dotenv

# own
pandas

